requirements.txt
asgiref==3.8.1
Django==5.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
tzdata==2024.1
